Move Platform URL helper to utils/checks instead of callbacks
Relocate PLATFORM_URL to utils and normalize_platform_uri to checks (next to check_file) so the file-resolution path no longer imports the heavy platform callbacks module just to rewrite a URL, enabling clean top-level imports.
