+++ docs/en/reference/utils/callbacks/platform.md
@@ -16,10 +16,6 @@ keywords: platform callbacks, training callbacks, console logging, YOLO11 traini
 
 <br><br><hr><br>
 
-## ::: ultralytics.utils.callbacks.platform.normalize_platform_uri
-
-<br><br><hr><br>
-
 ## ::: ultralytics.utils.callbacks.platform.resolve_platform_uri
 
 <br><br><hr><br>
